requirements.txt
torch>=2.0.0
torchvision
accelerate
scikit-learn>=1.0.0
surya-ocr>=0.4.0
Pillow>=10.0.0
numpy>=1.21.0
opencv-python-headless>=4.0.0
pypdfium2>=4.0.0
einops
timm
scipy
langdetect>=1.0.0
pyspellchecker>=0.7.0
transformers>=4.38.0
# detectron2 # Left commented as it requires specific build steps on Windows
